Fixed more tests
