+++ src/__tests__/package-sources/npmjs/stage-and-notify.lambda.test.ts
@@ -1,6 +1,6 @@
+import { GetObjectCommand, S3Client } from '@aws-sdk/client-s3';
 import { Context } from 'aws-lambda';
-import * as AWS from 'aws-sdk';
-import * as AWSMock from 'aws-sdk-mock';
+import { mockClient } from 'aws-sdk-client-mock';
 import * as nock from 'nock';
 import {
   ENV_DENY_LIST_BUCKET_NAME,
@@ -10,11 +10,10 @@ import {
   handler,
   PackageVersion,
 } from '../../../package-sources/npmjs/stage-and-notify.lambda';
+import { stringToStream } from '../../streams';
 
 const MOCK_DENY_LIST_BUCKET = 'deny-list-bucket-name';
 const MOCK_DENY_LIST_OBJECT = 'my-deny-list.json';
-
-type Response<T> = (err: AWS.AWSError | null, data?: T) => void;
 
 beforeEach(() => {
   process.env.BUCKET_NAME = 'foo';
@@ -34,13 +33,11 @@ test('ignores 404', async () => {
   const basePath = 'https://registry.npmjs.org';
   const uri = '/@pepperize/cdk-vpc/-/cdk-vpc-0.0.785.tgz';
 
-  AWSMock.mock(
-    'S3',
-    'getObject',
-    (_req: AWS.S3.GetObjectRequest, cb: Response<AWS.S3.GetObjectOutput>) => {
-      cb(null, { Body: JSON.stringify({}) });
-    }
-  );
+  const s3Mock = mockClient(S3Client);
+
+  s3Mock.on(GetObjectCommand).resolves({
+    Body: stringToStream(JSON.stringify({})),
+  });
 
   nock(basePath).get(uri).reply(404);
